% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sst_sim_fixed_ssd.R
\name{SstSimulatedFixedSsdBlock}
\alias{SstSimulatedFixedSsdBlock}
\title{One Block Model}
\usage{
SstSimulatedFixedSsdBlock(
  pid,
  n,
  m,
  SSD.b,
  dist.go,
  theta.go,
  dist.stop,
  theta.stop
)
}
\arguments{
\item{pid}{Participant id}

\item{n}{total number of trials (positive integer)}

\item{m}{total number of stops  (positive integer)}

\item{SSD.b}{one block stop signal delay}

\item{dist.go}{distribution of go trials (ExG or SW)}

\item{theta.go}{c(mu.go,sigma.go,tau.go)}

\item{dist.stop}{distribution of stop.trials (ExG or SW)}

\item{theta.stop}{c(mu.stop,sigma.stop,tau.stop)}
}
\value{
MAT11
}
\description{
Step(1): Do for one block and ExG/SW distribution
(for either Go/Stop)   OK
}
\examples{
## Example1
myblockSSTdata1 <- SstSimulatedFixedSsdBlock(pid="John.Smith", n=50, m=10,
                      SSD.b=200, dist.go="ExG", theta.go=c(400,60,30),
                      dist.stop="ExG", theta.stop=c(100,70,60))
myblockSSTdata1

## Example2
myblockSSTdata2 <- SstSimulatedFixedSsdBlock(pid="Jane.Smith", n=50, m=10,
                      SSD.b=200, dist.go="SW", theta.go=c(100,0.01,50),
                      dist.stop="SW", theta.stop=c(75,0.01,100))
myblockSSTdata2

## Example3         <-create error !
myblockSSTdata3 <- SstSimulatedFixedSsdBlock(pid="Jane.Smith", n=50, m=10,
                      SSD.b=200, dist.go ="LN", theta.go=c(100,0.01,50),
                      dist.stop="SW", theta.stop=c(75,0.01,100))
myblockSSTdata3

}
\references{

}
\author{

}
