% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sst_sim_trackmethod.R
\name{SstSimulatedTrackingMethodBlock}
\alias{SstSimulatedTrackingMethodBlock}
\alias{SstSimulatedTrackingMethod}
\title{Simulating SSRT data using tracking method}
\usage{
SstSimulatedTrackingMethodBlock(
  pid,
  n,
  m,
  SSD.b,
  dist.go,
  theta.go,
  dist.stop,
  theta.stop
)

SstSimulatedTrackingMethod(
  pid,
  n,
  m,
  SSD.b,
  dist.go,
  theta.go,
  dist.stop,
  theta.stop,
  block
)
}
\arguments{
\item{pid}{a vector of size b of Participant.id}

\item{n}{a vector of size b of total number of trials}

\item{m}{a vector of size b of total number of stops}

\item{SSD.b}{a vector of size b of starting stop signal delay}

\item{dist.go}{a vector of size b of distribution of go trials (ExG or SW)}

\item{theta.go}{c(mu.go,sigma.go,tau.go), a b*3 matrix}

\item{dist.stop}{a vector of size b of distribution of stop.trials (ExG or SW)}

\item{theta.stop}{c(mu.stop,sigma.stop,tau.stop), a b*3 matrix}

\item{block}{a block name vector of size b blocks}
}
\value{
Output a giant matrix with "sum(n)" rows and (7+1) columns
}
\description{
This function simulates b>=1 blocks of stop signal task trials for several participants using tracking method. Note that 'ExG' implies exponentially modified Gaussian and'SW' implies Shifted Wald distribution.
}
\examples{
#Example1
mySSTdata1<-SstSimulatedTrackingMethod(
    pid=c("John.Smith","Jane.McDonald","Jane.McDonald"),
    n=c(50,100,150), m=c(10,20,30),
    SSD.b=c(200,220,240), dist.go=c("ExG","ExG","ExG"),
    theta.go=as.matrix.data.frame(rbind(c(400,60,30),c(440,90,90),c(440,90,90))),
    dist.stop=c("ExG","ExG","ExG"),
    theta.stop=as.matrix.data.frame(rbind(c(100,70,60),c(120,80,70),c(120,80,70))),
    block=c(1,1,2))
mySSTdata1

#Example2
mySSTdata2 <- SstSimulatedTrackingMethod(
  pid=c("John.Smith","Jane.McDonald","Jane.McDonald"),
  n=c(50,100,150), m=c(10,20,30), SSD.b = c(200,220,240),
  dist.go=c("ExG","ExG","ExG"),
  theta.go=as.matrix.data.frame(rbind(c(400,60,30),c(440,90,90),c(440,90,90))),
  dist.stop=c("SW","SW","SW"),
  theta.stop=as.matrix.data.frame(rbind(c(75,0.01,100),c(75,0.01,100),c(75,0.01,100))),
  block=c(1,1,2))
mySSTdata2

# Example3  <--- produce error !
\dontrun{
mySSTdata3 <- SstSimulatedTrackingMethod(pid="John.Smith", n=c(1), m=c(50),
   SSD.b=c(200), dist.go=c("LN"),
   theta.go=as.matrix.data.frame(rbind(c(400,60,30))),
   dist.stop=c("SW"),
   theta.stop=as.matrix.data.frame(rbind(c(75,0.01,100))),
   block=c(1))
mySSTdata3
}

}
