% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sst_sim_trackmethod.R
\name{SstSimulatedTrackingMethodBlock}
\alias{SstSimulatedTrackingMethodBlock}
\title{One Block Model}
\usage{
SstSimulatedTrackingMethodBlock(
  pid,
  n,
  m,
  SSD.b,
  dist.go,
  theta.go,
  dist.stop,
  theta.stop
)
}
\arguments{
\item{pid}{Participant.id}

\item{n}{total number of trials (positive integer)}

\item{m}{total number of stops  (positive integer)}

\item{SSD.b}{one block starting stop signal delay}

\item{dist.go}{distribution of go trials    (ExG or SW)}

\item{theta.go}{c(mu.go,sigma.go,tau.go)}

\item{dist.stop}{distribution of stop.trials  (ExG or SW)}

\item{theta.stop}{c(mu.stop,sigma.stop,tau.stop)}
}
\value{
MAT22
}
\description{
Step(1): Do for one block and ExG/SW distribution
(for either Go/Stop)   OK
}
\references{

}
\author{

}
