% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simssfixed.R
\name{simssfixed0}
\alias{simssfixed0}
\alias{simssfixed}
\title{Simulatng SSRT data using fixed SSD methods}
\usage{
simssfixed0(pid, n, m, SSD.b, dist.go, theta.go, dist.stop, theta.stop)

simssfixed(pid, block, n, m, SSD.b, dist.go, theta.go, dist.stop, theta.stop)
}
\arguments{
\item{pid}{a vector of size b of Participant.id}

\item{n}{a vector of size b of total number of trials}

\item{m}{a vector of size b of total number of stops}

\item{SSD.b}{a vector of size b of stop signal delay}

\item{dist.go}{a vector of size b of distribution of go trials (ExG or SW)}

\item{theta.go}{c(mu.go, sigma.go, tau.go), a b*3 matrix}

\item{dist.stop}{a vector of size b of distribution of stop.trials (ExG or SW)}

\item{theta.stop}{c(mu.stop, sigma.stop, tau.stop), a b*3 matrix}

\item{block}{a block name vector of size b blocks}
}
\value{
M11
Output: a giant matrix with "sum(n)" rows and (7+1) columns
}
\description{
This function simulates b>=1 blocks of stop signal task trials for several participants using fixed SSD method.
}
\examples{
## Example1
mySSTdata1 <- simssfixed(
 pid = c("John.Smith","Jane.McDonald","Jane.McDonald"),
 n = c(50,100,150), m=c(10,20,30), SSD.b=c(200,220,240),
 dist.go=c("ExG","ExG","ExG"),
 theta.go=as.matrix(rbind(c(400,60,30),c(440,90,90),c(440,90,90))),
 dist.stop=c("ExG","ExG","ExG"),
 theta.stop=as.matrix(rbind(c(100,70,60),c(120,80,70),c(120,80,70))),
 block=c(1,1,2))
mySSTdata1

## Example2
mySSTdata2 <- simssfixed(
  pid=c("John.Smith","Jane.McDonald","Jane.McDonald"),
  n=c(50,100,150), m=c(10,20,30), SSD.b=c(200,220,240),
  dist.go=c("ExG","ExG","ExG"),
  theta.go=as.matrix(rbind(c(400,60,30),c(440,90,90),c(440,90,90))),
  dist.stop=c("SW","SW","SW"),
  theta.stop=as.matrix(rbind(c(75,0.01,100),c(75,0.01,100),c(75,0.01,100))),
  block=c(1,1,2))
mySSTdata2

## Example3
\dontrun{ ## this expected to produce an error
mySSTdata3 <- simssfixed(
    pid=c("John.Smith"),
    n=c(50), m=c(10), SSD.b=c(200),
    dist.go=c("LN"),
    theta.go=as.matrix.data.frame(rbind(c(400,60,30))),
    dist.stop=c("SW"),
    theta.stop=as.matrix.data.frame(rbind(c(75,0.01,100))),
    block=c(1))
mySSTdata3
}
}
\references{
Gordon D. Logan. On the Ability to Inhibit Thought and Action: A User's Guide to the Stop Signal Paradigm. In D. Dagenbach, & T.H. Carr (Eds.), Inhibitory Process in Attention, Memory and Language. San Diego: Academic Press, 1994.
Dora Matzke, Jonathon Love, Thomas V. Wiecki, Scott D. Brown, and et al. Release the BEESTS: Bayesian Estimation of Ex-Gaussian Stop Signal Reaction Times Distributions. Frontiers in Psychology, 4: Article 918, 2013.
Royce Andres, F. Xavier Alario, and Leendert Van Maanen. The Shifted Wald Distribution for Response Time Data Analysis. Psychological Methods, 21(3): 309-327, 2016.
Andrew Heathcote. RTSYS: A DOS Application for the Analysis of Reaction Time Data, Behavior Research Methods, Instruments, & Computers, 28(3): 427-445, 1996.
Mohsen Soltanifar. Stop Signal Reaction Times: New Estimations with Longitudinal, Bayesian and Time  Series based Methods, PhD Dissertation, Biostatistics Division, Dalla Lana School of Public Health, University of Toronto, Toronto, Canada, 2020.
}
